Added some thread safety

+++ McRider.Common/Services/FileCacheService.cs
@@ -1,15 +1,21 @@
-﻿using Newtonsoft.Json;
+﻿using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
 using System.Reflection;
 
 namespace McRider.Common.Services;
 
 public class FileCacheService
 {
+    private ILogger _logger;
     private readonly string _cacheDirectory;
+    private readonly JsonSerializerSettings settings;
+    private readonly SemaphoreSlim semaphore = new SemaphoreSlim(1, 2);
+
     public TimeSpan? DefaultCacheDuration { get; set; }
 
-    public FileCacheService()
+    public FileCacheService(ILogger<FileCacheService> logger = null)
     {
+        _logger = logger;
         var currentAssemblyName = string.Join(".", (Assembly.GetExecutingAssembly()?.GetName()?.Name ?? "files.").Split('.').SkipLast(1));
         _cacheDirectory = Path.Combine(
             Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments),
@@ -18,37 +24,81 @@ public class FileCacheService
 
         if (!Directory.Exists(_cacheDirectory))
             Directory.CreateDirectory(_cacheDirectory);
+
+        settings = new JsonSerializerSettings
+        {
+            PreserveReferencesHandling = PreserveReferencesHandling.Objects, // Preserve object references
+            ReferenceLoopHandling = ReferenceLoopHandling.Serialize, // Handle reference loops
+            Formatting = Formatting.Indented // Optional, for pretty-printing JSON
+        };
     }
 
     public string Get(string key, TimeSpan? cacheDuration = null)
     {
         string filePath = GetCacheFilePath(key);
-        if (File.Exists(filePath))
-        {
-            cacheDuration ??= DefaultCacheDuration;
-            if (cacheDuration == null)
-                return File.ReadAllText(filePath);
+        if (!File.Exists(filePath))
+            return null; // Cache does not exist
 
-            var cacheExpiry = File.GetLastWriteTime(filePath).Add(cacheDuration.Value);
-            if (DateTime.UtcNow > cacheExpiry)
-                return null;
+
+        try
+        {
+            semaphore.Wait();
+
+            cacheDuration ??= DefaultCacheDuration;
+            if (cacheDuration != null)
+            {
+                var cacheExpiry = File.GetLastWriteTime(filePath).Add(cacheDuration.Value);
+                if (DateTime.UtcNow > cacheExpiry)
+                    return null; // Cache has expired
+            }
 
             return File.ReadAllText(filePath);
         }
-
-        return null;
+        catch (Exception e)
+        {
+            _logger?.LogError(e, "Error while reading file: {0}", filePath);
+            return null;
+        }
+        finally
+        {
+            semaphore.Release(); // Release the semaphore
+        }
     }
 
     public void Set(string key, string data)
     {
         string filePath = GetCacheFilePath(key);
-        File.WriteAllText(filePath, data);
+        try
+        {
+            semaphore.Wait();
+            File.WriteAllText(filePath, data);
+        }
+        catch (Exception e)
+        {
+            _logger?.LogError(e, "Error while writing to cache file: {0}", filePath);
+        }
+        finally
+        {
+            semaphore.Release();
+        }
     }
 
     public async Task SetAsync<T>(string key, T data)
     {
         string filePath = GetCacheFilePath(key);
-        await File.WriteAllTextAsync(filePath, JsonConvert.SerializeObject(data, Formatting.Indented));
+        try
+        {
+            await semaphore.WaitAsync();
+            await File.WriteAllTextAsync(filePath, JsonConvert.SerializeObject(data, settings));
+        }
+        catch (Exception e)
+        {
+            _logger?.LogError(e, "Error while writing to cache file: {0}", filePath);
+        }
+        finally
+        {
+            semaphore.Release();
+        }
     }
 
     public T Get<T>(string key, Func<T> readFunc, TimeSpan? cacheDuration = null)
@@ -59,12 +109,12 @@ public class FileCacheService
         {
             T freshData = readFunc();
 
-            Set(key, JsonConvert.SerializeObject(freshData));
+            _ = SetAsync(key, freshData);
 
             return freshData;
         }
 
-        return JsonConvert.DeserializeObject<T>(cachedData);
+        return JsonConvert.DeserializeObject<T>(cachedData, settings);
     }
 
     public async Task<T> GetAsync<T>(string key, Func<Task<T>> readFunc, TimeSpan? cacheDuration = null)
@@ -80,14 +130,26 @@ public class FileCacheService
             return freshData;
         }
 
-        return JsonConvert.DeserializeObject<T>(cachedData);
+        return JsonConvert.DeserializeObject<T>(cachedData, settings);
     }
 
     public void Remove(string key)
     {
         string filePath = GetCacheFilePath(key);
-        if (File.Exists(filePath))
-            File.Delete(filePath);
+        try
+        {
+            semaphore.Wait();
+            if (File.Exists(filePath)) File.Delete(filePath);
+        }
+        catch (Exception e)
+        {
+            _logger?.LogError(e, "Error while deleting cache file: {0}", filePath);
+        }
+        finally
+        {
+            semaphore.Release();
+        }
+
     }
 
     private string GetCacheFilePath(string key)
